requirements.txt
click==6.6
docker-py==1.8.1
